fix: Sanitised input and removed redundant checks
